requirements.txt
Django==4.2.7
PyPDF2==3.0.1
python-docx==1.1.0